Refactored my suggestion page
